move scylla db to databese from state indexer
