+++ database/Cargo.toml
@@ -6,8 +6,13 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.70"
 async-trait = "0.1.66"
+borsh = "0.10.2"
+futures = "0.3.5"
+hex = "0.4.3"
 prettytable-rs = { version = "0.10", optional = true }
-scylla = "0.9.0"
+num-bigint = { version = "0.3", optional = true }
+num-traits = { version = "0.2.15", optional = true }
+scylla = { version = "0.9.0", optional = true }
 tracing = "0.1.34"
 uuid = { version = "1.3.0", optional = true }
 
@@ -15,7 +20,9 @@ readnode-primitives = { path = "../readnode-primitives" }
 
 near-primitives = "0.17.0"
 near-crypto = "0.17.0"
+near-indexer-primitives = "0.17.0"
 
 [features]
-scylla_db_tracing = ["dep:prettytable-rs", "dep:uuid"]
+scylla_db_tracing = ["dep:prettytable-rs", "dep:uuid", "dep:scylla", "dep:num-bigint", "dep:num-traits"]
+scylla_db = ["dep:scylla", "dep:num-bigint", "dep:num-traits"]
 account_access_keys = []
