+++ database/src/lib.rs
@@ -1,5 +1,9 @@
 mod base;
-mod scylladb;
-
 pub use crate::base::RpcDbManager;
-pub use crate::scylladb::ScyllaStorageManager;
+pub use crate::base::StateIndexerDbManager;
+mod scylladb;
+pub use crate::scylladb::base::ScyllaStorageManager;
+
+pub mod primitives;
+pub mod rpc_server;
+pub mod state_indexer;
