+++ database/src/primitives.rs
@@ -1,0 +1,3 @@
+pub type ShardId = u64;
+pub type HeightIncluded = u64;
+pub type ChunkHash = String;
