+++ database/src/scylladb/mod.rs
@@ -1,0 +1,3 @@
+pub mod base;
+pub mod rpc_server;
+pub mod state_indexer;
