+++ rpc-server/Cargo.toml
@@ -22,11 +22,8 @@ http = "0.2.8"
 jsonrpc-v2 = { git = "https://github.com/kobayurii/jsonrpc-v2", rev = "95e7b1d2567ae841163af212a3f25abb6862becb" }
 lazy_static = "1.4.0"
 lru = "0.11.1"
-num-bigint = "0.3"
-num-traits = "0.2.15"
 paste = "1.0.14"
 prometheus = "0.13.1"
-scylla = "0.9.0"
 serde = { version = "1.0.145", features = ["derive"] }
 serde_json = "1.0.85"
 thiserror = "1.0.40"
@@ -65,9 +62,10 @@ near-vm-errors = "0.17.0"
 
 
 [features]
-default = ["submit_tx_methods"]
+default = ["submit_tx_methods", "scylla_db"]
 submit_tx_methods = []
 tracing-instrumentation = []
-scylla_db_tracing = ["database/scylla_db_tracing"]
+scylla_db_tracing = ["database/scylla_db_tracing", "database/scylla_db"]
+scylla_db = ["database/scylla_db"]
 shadow_data_consistency = ["dep:assert-json-diff"]
 account_access_keys = ["database/account_access_keys"]
