+++ rpc-server/src/config.rs
@@ -1,7 +1,5 @@
 use crate::modules::blocks::CacheBlock;
-use crate::storage;
 use clap::Parser;
-use database::ScyllaStorageManager;
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -14,27 +12,17 @@ pub struct Opts {
     #[clap(long, env = "AWS_BUCKET_NAME")]
     pub s3_bucket_name: String,
 
-    // Scylla db url
+    // DB url
     #[clap(long, env)]
-    pub scylla_url: String,
+    pub database_url: String,
 
-    /// ScyllaDB user(login)
+    /// DB user(login)
     #[clap(long, env)]
-    pub scylla_user: Option<String>,
+    pub database_user: Option<String>,
 
-    /// ScyllaDB password
+    ///DB password
     #[clap(long, env)]
-    pub scylla_password: Option<String>,
-
-    /// ScyllaDB preferred DataCenter
-    /// Accepts the DC name of the ScyllaDB to filter the connection to that DC only (preferrably).
-    /// If you connect to multi-DC cluter, you might experience big latencies while working with the DB. This is due to the fact that ScyllaDB driver tries to connect to any of the nodes in the cluster disregarding of the location of the DC. This option allows to filter the connection to the DC you need. Example: "DC1" where DC1 is located in the same region as the application.
-    #[clap(long, env)]
-    pub scylla_preferred_dc: Option<String>,
-
-    /// ScyllaDB keepalive interval
-    #[clap(long, env, default_value = "60")]
-    pub scylla_keepalive_interval: u64,
+    pub database_password: Option<String>,
 
     // AWS access key id
     #[clap(long, env = "AWS_ACCESS_KEY_ID")]
@@ -51,16 +39,6 @@ pub struct Opts {
     // AWS default region
     #[clap(long, env, default_value = "8000")]
     pub server_port: u16,
-
-    /// Max retry count for ScyllaDB if `strict_mode` is `false`
-    #[clap(long, default_value = "2", env)]
-    pub max_retry: u8,
-
-    /// Attempts to store data in the database should be infinite to ensure no data is missing.
-    /// Disable it to perform a limited write attempts (`max_retry`)
-    /// before skipping giving up and moving to the next piece of data
-    #[clap(long, default_value = "false", env)]
-    pub strict_mode: bool,
 
     /// Max gas burnt for contract function call
     /// Default value is 300_000_000_000_000
@@ -112,20 +90,6 @@ impl Opts {
             .start_block_height(start_block_height)
             .build()
             .expect("Failed to build LakeConfig"))
-    }
-
-    pub async fn prepare_db_manager(&self) -> anyhow::Result<impl database::RpcDbManager> {
-        let db_manager = *storage::ScyllaDBManager::new(
-            &self.scylla_url,
-            self.scylla_user.as_deref(),
-            self.scylla_password.as_deref(),
-            self.scylla_preferred_dc.as_deref(),
-            Some(self.scylla_keepalive_interval),
-            self.max_retry,
-            self.strict_mode,
-        )
-        .await?;
-        Ok(db_manager)
     }
 }
