+++ state-indexer/Cargo.toml
@@ -19,11 +19,8 @@ http = "0.2"
 humantime = "2.1.0"
 itertools = "0.10.0"
 lazy_static = "1.4.0"
-num-bigint = "0.3"
-num-traits = "0.2.15"
 openssl-probe = "0.1.5"
 prometheus = "0.13.1"
-scylla = "0.9.0"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.55"
 tokio = { version = "1.1", features = [
@@ -53,10 +50,12 @@ database = { path = "../database" }
 
 near-primitives-core = "0.17.0"
 near-indexer-primitives = "0.17.0"
-near-jsonrpc-client = { git = "https://github.com/khorolets/near-jsonrpc-client-rs", rev = "0b6ad111307202a37028ef300a75a41f2a7947c3" }
+near-jsonrpc-client = "0.6.0"
 near-lake-framework = "0.7.2"
 
 [features]
+default = ["scylla_db"]
 tracing-instrumentation = []
-scylla_db_tracing = ["database/scylla_db_tracing"]
-account_access_keys = []
+scylla_db_tracing = ["database/scylla_db_tracing", "database/scylla_db"]
+scylla_db = ["database/scylla_db"]
+account_access_keys = ["database/account_access_keys"]
